chore: rename pages dir files
