+++ src/pages/_app.page.tsx
@@ -1,4 +1,4 @@
-import '@/styles/globals.css'
+import "../styles/globals.css"
 import type { AppProps } from 'next/app'
 
 export default function App({ Component, pageProps }: AppProps) {
