+++ src/pages/index.page.tsx
@@ -1,0 +1,17 @@
+import Head from 'next/head'
+
+export default function Home() {
+  return (
+    <>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <main>
+        <p className="text-2xl">Hello world!!</p>
+      </main>
+    </>
+  )
+}
